Add files via upload
Added more functionality to the main window. when the game launches all the properties load in with their respective colours (barring opportunity knocks, and pot luck cards). there are also new windows for adding players into the game and there is an option to have a time-restricted game(part functionality) and a normal game (not functional yet). other housekeeping methods have been implemented such as Player Pieces. The Human class will be changed to the player and will be abstract once the human side of the game is functional. there is a countdown timer which allows the players to see how much time that has left for the game.